+++ vendor_modules/dnsquery/lib/puppet_x/voxpupuli/dnsquery/util.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module PuppetX
+  module Voxpupuli
+    module Dnsquery
+      module Utils
+        def self.resolver(config_info = nil)
+          if config_info
+            Resolv::DNS.new(config_info.transform_keys(&:to_sym))
+          else
+            Resolv::DNS.new
+          end
+        end
+      end
+    end
+  end
+end