+++ vendor_modules/dnsquery/spec/functions/dnsquery_a_spec.rb
@@ -10,6 +10,24 @@ describe 'dnsquery::a' do
     expect(results).to all(match(%r{^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$}))
   end
 
+  it 'returns a list of IPv4 addresses when doing a lookup with updated namesrver' do
+    results = subject.execute('google.com', { 'nameserver' => '8.8.8.8' })
+    expect(results).to be_a Array
+    expect(results).to all(match(%r{^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$}))
+  end
+
+  it 'returns a list of IPv4 addresses when doing a lookup with updated ndots' do
+    results = subject.execute('google.com', { 'nameserver' => '8.8.8.8', 'ndots' => 1 })
+    expect(results).to be_a Array
+    expect(results).to all(match(%r{^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$}))
+  end
+
+  it 'returns a list of IPv4 addresses when doing a lookup with updated search' do
+    results = subject.execute('google', { 'nameserver' => '8.8.8.8', 'search' => ['com'] })
+    expect(results).to be_a Array
+    expect(results).to all(match(%r{^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$}))
+  end
+
   it 'returns lambda value if result is empty' do
     is_expected.to(
       run.