+++ vendor_modules/dnsquery/spec/functions/dnsquery_mx_spec.rb
@@ -17,6 +17,36 @@ describe 'dnsquery::mx' do
     end
   end
 
+  it 'returns a list of MX records when doing a lookup with update nameserver' do
+    results = subject.execute('google.com', { 'nameserver' => ['8.8.8.8'] })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['preference']).to be_a Integer
+      expect(res['exchange']).to be_a String
+    end
+  end
+
+  it 'returns a list of MX records when doing a lookup with update ndots' do
+    results = subject.execute('google.com', { 'nameserver' => '8.8.8.8', 'ndots' => 1 })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['preference']).to be_a Integer
+      expect(res['exchange']).to be_a String
+    end
+  end
+
+  it 'returns a list of MX records when doing a lookup with update search' do
+    results = subject.execute('google', { 'nameserver' => '8.8.8.8', 'search' => ['com'] })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['preference']).to be_a Integer
+      expect(res['exchange']).to be_a String
+    end
+  end
+
   it 'returns lambda value if result is empty' do
     is_expected.to(
       run.