+++ vendor_modules/dnsquery/spec/functions/dnsquery_srv_spec.rb
@@ -21,6 +21,42 @@ describe 'dnsquery::srv' do
     end
   end
 
+  it 'returns a list of SRV records when doing a lookup with update nameserver' do
+    results = subject.execute('_spotify-client._tcp.spotify.com', { 'nameserver' => ['8.8.8.8'] })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['priority']).to be_a Integer
+      expect(res['weight']).to be_a Integer
+      expect(res['port']).to be_a Integer
+      expect(res['target']).to be_a String
+    end
+  end
+
+  it 'returns a list of SRV records when doing a lookup with update ndots' do
+    results = subject.execute('_spotify-client._tcp.spotify.com', { 'nameserver' => '8.8.8.8', 'ndots' => 1 })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['priority']).to be_a Integer
+      expect(res['weight']).to be_a Integer
+      expect(res['port']).to be_a Integer
+      expect(res['target']).to be_a String
+    end
+  end
+
+  it 'returns a list of SRV records when doing a lookup with update search' do
+    results = subject.execute('_spotify-client._tcp.spotify', { 'nameserver' => '8.8.8.8', 'search' => ['com'] })
+    expect(results).to be_a Array
+    expect(results).to all(be_a(Hash))
+    results.each do |res|
+      expect(res['priority']).to be_a Integer
+      expect(res['weight']).to be_a Integer
+      expect(res['port']).to be_a Integer
+      expect(res['target']).to be_a String
+    end
+  end
+
   it 'returns lambda value if result is empty' do
     is_expected.to(
       run.